+++ src/Form/Element/DateTime.php
@@ -6,10 +6,11 @@ use Exception;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Log;
 use SleepingOwl\Admin\Traits\DatePicker;
+use SleepingOwl\Admin\Traits\DateFormat;
 
 class DateTime extends NamedFormElement
 {
-    use DatePicker;
+    use DatePicker, DateFormat;
 
     /**
      * @var string
@@ -31,45 +32,6 @@ class DateTime extends NamedFormElement
      */
     protected $view = 'form.element.datetime';
 
-    /**
-     * @return string
-     */
-    public function getFormat()
-    {
-        return $this->format;
-    }
-
-    /**
-     * @return string
-     */
-    public function getTimezone()
-    {
-        return $this->timezone ?: config('sleeping_owl.timezone');
-    }
-
-    /**
-     * @param string|null $format
-     *
-     * @return $this
-     */
-    public function setFormat($format)
-    {
-        $this->format = $format;
-
-        return $this;
-    }
-
-    /**
-     * @param string $timezone
-     *
-     * @return $this
-     */
-    public function setTimezone($timezone)
-    {
-        $this->timezone = $timezone;
-
-        return $this;
-    }
 
     /**
      * @return $this|NamedFormElement|mixed|null|string
@@ -167,13 +129,12 @@ class DateTime extends NamedFormElement
         if (empty($date)) {
             return;
         }
-
         if (! $date instanceof Carbon) {
             try {
                 $date = Carbon::parse($date);
             } catch (\Exception $e) {
                 try {
-                    $date = Carbon::createFromFormat($this->getPickerFormat(), $date);
+                  $date = Carbon::createFromFormat($this->getPickerFormat(), $date);
                 } catch (\Exception $e) {
                     Log::error('Unable to parse date!', [
                         'format' => $this->getPickerFormat(),
@@ -186,33 +147,6 @@ class DateTime extends NamedFormElement
             }
         }
 
-        return $date->timezone($this->getTimezone())->format($this->getFormat());
+        return $date->timezone($this->getTimezone())->format($this->getPickerFormat());
     }
-
-    // protected function parseValue($value)
-    // {
-    //     try {
-    //         $time = Carbon::parse($value);
-    //     } catch (Exception $e) {
-    //         Log::info('unable to parse date, re-trying with given format', [
-    //             'exception' => $e,
-    //             'date' => $value,
-    //         ]);
-    //         try {
-    //             $time = Carbon::createFromFormat($this->getPickerFormat(), $value);
-    //         } catch (Exception $e) {
-    //             Log::error('unable to parse date!', [
-    //                 'exception' => $e,
-    //                 'pickerFormat' => $this->getPickerFormat(),
-    //                 'date' => $value,
-    //             ]);
-    //
-    //             return;
-    //         }
-    //     }
-    //
-    //     return $time->timezone($this->getTimezone())->format(
-    //         $this->getPickerFormat()
-    //     );
-    // }
 }