+++ src/Traits/DateFormat.php
@@ -12,7 +12,6 @@ trait DateFormat
         if (is_null($this->format)) {
             $this->format = config('sleeping_owl.datetimeFormat');
         }
-
         return $this->format;
     }